upgradeHandler for dig v2
